requirements.txt
# Olfex (formerly PriceHunter) Production Requirements

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP & Network
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Processing
beautifulsoup4>=4.12.0
rapidfuzz>=3.5.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Cache & Message Queue
redis>=5.0.0

# Scheduling
apscheduler>=3.10.0

# Image Processing
Pillow>=10.0.0

# Environment & Config
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Logging & Monitoring
python-json-logger>=2.0.0

# Security
cryptography>=41.0.0

# Type Hints
typing-extensions>=4.8.0

# Utilities
click>=8.1.0
